Resource rework (#56)
* trying resourceManager design

* ResourceManager

* intermediate not working

* fix warning

* fix ResourceManager template deduction

* fixed tests

using ResourceManager in asteroids for testing

* replaced typedResourceCache with new ResourceManager

* click events for menu buttons

* Some idea for FramedSprite

* FramedSprite with callback

delayed entity removal via event

* prevent missing sound resources crashing the game

* Build fix for gcc

added cmake build presets

---------

+++ apps/asteroids/systems/RenderSystem.cpp
@@ -41,7 +41,7 @@ static void renderSDL(sdl::Renderer& renderer, const pg::BoundingSphere& bs, con
 
         circle_points.push_back(pg::vec_cast<int>(p));
     }
-    renderer.drawLines(std::bit_cast<SDL_Point*>(circle_points.data()), circle_points.size());
+    renderer.drawLines(std::bit_cast<SDL_Point*>(circle_points.data()), static_cast<int>(circle_points.size()));
 }
 
 void asteroids::RenderSystem::handle(const pg::game::FrameStamp&)
